estudos/projetos-python/basico01/pyproject.toml
[project]
name = "basico01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "evolutionapi>=0.0.6",
    "flask>=3.1.0",
    "google-generativeai>=0.8.3",
    "numpy>=2.2.1",
    "pygame>=2.6.1",
    "python-dotenv>=1.0.1",
    "requests-toolbelt>=1.0.0",
    "streamlit>=1.41.1",
]
